requirements.txt
torch
torchvision
Pillow
opencv-python
